+++ cyto_dl/nn/vits/__init__.py
@@ -1,3 +1,4 @@
 from .cross_mae import CrossMAE_Decoder
 from .mae import MAE_Decoder, MAE_Encoder, MAE_ViT
 from .seg import Seg_ViT, SuperresDecoder
+from .utils import take_indexes
