+++ cyto_dl/nn/vits/blocks/__init__.py
@@ -1,1 +1,3 @@
 from .cross_attention import CrossAttention, CrossAttentionBlock, CrossSelfBlock, Mlp
+from .intermediate_weigher import IntermediateWeigher
+from .patchify import Patchify
