+++ cyto_dl/nn/vits/cross_mae.py
@@ -94,6 +94,8 @@ class CrossMAE_Decoder(torch.nn.Module):
         trunc_normal_(self.pos_embedding, std=0.02)
 
     def forward(self, features, forward_indexes, backward_indexes):
+        # HACK TODO allow usage of multiple intermediate feature weights, this works when decoder is 0 layers
+        features = features.squeeze(0)
         T, B, C = features.shape
         # we could do cross attention between decoder_dim queries and encoder_dim features, but it seems to work fine having both at decoder_dim for now
         features = self.projection_norm(self.projection(features))
